Refactor video cover utils
