+++ registry/lib/components/utils/view-cover/types.ts
@@ -1,0 +1,1 @@
+export type CoverDownloadType = 'jpg'
